[refactor] #107 swagger 문서 위치 이동
