+++ src/main/java/nova/backend/domain/cafe/controller/StaffCafeController.java
@@ -1,9 +1,10 @@
 package nova.backend.domain.cafe.controller;
 
 import lombok.RequiredArgsConstructor;
+import nova.backend.domain.cafe.controller.api.StaffCafeApi;
 import nova.backend.domain.cafe.dto.response.CafeMyListItemDTO;
 import nova.backend.domain.cafe.dto.response.CafeOwnerSelectedResponseDTO;
-import nova.backend.domain.cafe.service.CafeSelectionService;
+import nova.backend.domain.cafe.service.StaffCafeService;
 import nova.backend.global.auth.CustomUserDetails;
 import nova.backend.global.common.SuccessResponse;
 import org.springframework.http.ResponseEntity;
@@ -15,15 +16,15 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/staff/cafes")
-public class CafeSelectionController implements CafeSelectionApi{
-    private final CafeSelectionService cafeSelectionService;
+public class StaffCafeController implements StaffCafeApi {
+    private final StaffCafeService staffCafeService;
 
     @PutMapping("/{cafeId}/selected")
     public ResponseEntity<SuccessResponse<?>> selectCafe(
             @AuthenticationPrincipal CustomUserDetails userDetails,
             @PathVariable Long cafeId
     ) {
-        cafeSelectionService.selectCafe(userDetails.getUserId(), cafeId);
+        staffCafeService.selectCafe(userDetails.getUserId(), cafeId);
         return SuccessResponse.ok("선택된 카페 설정 완료");
     }
 
@@ -31,7 +32,7 @@ public class CafeSelectionController implements CafeSelectionApi{
     public ResponseEntity<SuccessResponse<?>> getMyCafes(
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
-        List<CafeMyListItemDTO> response = cafeSelectionService.getMyCafes(
+        List<CafeMyListItemDTO> response = staffCafeService.getMyCafes(
                 userDetails.getUserId(),
                 userDetails.getSelectedCafeId()
         );
@@ -42,7 +43,7 @@ public class CafeSelectionController implements CafeSelectionApi{
     public ResponseEntity<SuccessResponse<?>> getSelectedCafe(
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
-        CafeOwnerSelectedResponseDTO response = cafeSelectionService.getSelectedCafe(userDetails.getSelectedCafeId());
+        CafeOwnerSelectedResponseDTO response = staffCafeService.getSelectedCafe(userDetails.getSelectedCafeId());
         return SuccessResponse.ok(response);
     }
 }
