+++ src/main/java/nova/backend/domain/cafe/controller/api/OwnerCafeApi.java
@@ -1,4 +1,4 @@
-package nova.backend.domain.cafe.controller;
+package nova.backend.domain.cafe.controller.api;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
