+++ src/main/java/nova/backend/domain/stamp/controller/api/StampApi.java
@@ -1,14 +1,12 @@
-package nova.backend.domain.stamp.controller;
+package nova.backend.domain.stamp.controller.api;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import nova.backend.domain.stamp.dto.request.StampAccumulateRequestDTO;
 import nova.backend.domain.stamp.schema.StampHistoryListSuccessResponse;
 import nova.backend.global.common.SuccessResponse;
 import org.springframework.http.ResponseEntity;
